+++ src/widget/form/index.ts
@@ -1,1 +1,2 @@
 export { AuthForm } from './AuthForm';
+export { SignForm } from './SignForm';
